Move Debug.cs from Common to CoreLib
Stops VS 2017 from complaining about duplicate Debug.cs, and preps us for sharing with CoreLib.
