--- src/Common/src/System/Diagnostics/Debug.cs
@@ -1,225 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-#define DEBUG // Do not remove this, it is needed to retain calls to these conditional methods in release builds
-
-namespace System.Diagnostics
-{
-    /// <summary>
-    /// Provides a set of properties and methods for debugging code.
-    /// </summary>
-    public static partial class Debug
-    {
-        private static readonly object s_ForLock = new Object();
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Assert(bool condition)
-        {
-            Assert(condition, string.Empty, string.Empty);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Assert(bool condition, string message)
-        {
-            Assert(condition, message, string.Empty);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Assert(bool condition, string message, string detailMessage)
-        {
-            if (!condition)
-            {
-                string stackTrace;
-
-                try
-                {
-                    stackTrace = Environment.StackTrace;
-                }
-                catch
-                {
-                    stackTrace = "";
-                }
-
-                WriteAssert(stackTrace, message, detailMessage);
-                s_logger.ShowAssertDialog(stackTrace, message, detailMessage);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Fail(string message)
-        {
-            Assert(false, message, string.Empty);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Fail(string message, string detailMessage)
-        {
-            Assert(false, message, detailMessage);
-        }
-
-        private static void WriteAssert(string stackTrace, string message, string detailMessage)
-        {
-            string assertMessage = SR.DebugAssertBanner + Environment.NewLine
-                                   + SR.DebugAssertShortMessage + Environment.NewLine
-                                   + message + Environment.NewLine
-                                   + SR.DebugAssertLongMessage + Environment.NewLine
-                                   + detailMessage + Environment.NewLine
-                                   + stackTrace;
-
-            WriteLine(assertMessage);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Assert(bool condition, string message, string detailMessageFormat, params object[] args)
-        {
-            Assert(condition, message, string.Format(detailMessageFormat, args));
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLine(string message)
-        {
-            Write(message + Environment.NewLine);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Write(string message)
-        {
-            s_logger.WriteCore(message);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLine(object value)
-        {
-            WriteLine((value == null) ? string.Empty : value.ToString());
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLine(object value, string category)
-        {
-            WriteLine((value == null) ? string.Empty : value.ToString(), category);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLine(string format, params object[] args)
-        {
-            WriteLine(string.Format(null, format, args));
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLine(string message, string category)
-        {
-            if (category == null)
-            {
-                WriteLine(message);
-            }
-            else
-            {
-                WriteLine(category + ":" + ((message == null) ? string.Empty : message));
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Write(object value)
-        {
-            Write((value == null) ? string.Empty : value.ToString());
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Write(string message, string category)
-        {
-            if (category == null)
-            {
-                Write(message);
-            }
-            else
-            {
-                Write(category + ":" + ((message == null) ? string.Empty : message));
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void Write(object value, string category)
-        {
-            Write((value == null) ? string.Empty : value.ToString(), category);
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteIf(bool condition, string message)
-        {
-            if (condition)
-            {
-                Write(message);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteIf(bool condition, object value)
-        {
-            if (condition)
-            {
-                Write(value);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteIf(bool condition, string message, string category)
-        {
-            if (condition)
-            {
-                Write(message, category);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteIf(bool condition, object value, string category)
-        {
-            if (condition)
-            {
-                Write(value, category);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLineIf(bool condition, object value)
-        {
-            if (condition)
-            {
-                WriteLine(value);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLineIf(bool condition, object value, string category)
-        {
-            if (condition)
-            {
-                WriteLine(value, category);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLineIf(bool condition, string value)
-        {
-            if (condition)
-            {
-                WriteLine(value);
-            }
-        }
-
-        [System.Diagnostics.Conditional("DEBUG")]
-        public static void WriteLineIf(bool condition, string value, string category)
-        {
-            if (condition)
-            {
-                WriteLine(value, category);
-            }
-        }
-
-        internal interface IDebugLogger
-        {
-            void ShowAssertDialog(string stackTrace, string message, string detailMessage);
-            void WriteCore(string message);
-        }
-    }
-}
